Access types and checkAccess removal
