+++ src/backgroundScript.ts
@@ -18,7 +18,7 @@ import {
 import { initBackgroundSignChannel } from './dApps/SignChannels/backgroundSignChannel';
 import { initBackgroundSignRawChannel } from './dApps/SignRawChannels/backgroundSignRawChannel';
 import { backgroundChallengeChannel } from './channels/ChallengeChannels/backgroundChallengeChannel';
-import { initBackgroundAccessChannel } from './dApps/AccessChannels/browserAccessChannels';
+import { initBackgroundAccessChannel } from './dApps/AccessChannels/backgroundAccessChannels';
 import { initBackgroundGenesisHashChannel } from './dApps/genesisHashChannel/initBackgroundGenesisHashChannel';
 import { initBackgroundCredentialChannel } from './channels/CredentialChannels/backgroundCredentialChannel';
 import { initBackgroundSignDidChannel } from './channels/SignDidChannels/backgroundSignDidChannel';
