+++ src/dApps/AccessChannels/types.ts
@@ -1,0 +1,6 @@
+export interface AccessInput {
+  dAppName: string;
+  origin: string;
+}
+
+export type AccessOutput = boolean;
