+++ src/dApps/SignRawChannels/contentSignRawChannel.ts
@@ -1,5 +1,5 @@
 import { BrowserChannel } from '../../channels/base/BrowserChannel/BrowserChannel';
-import { checkAccess } from '../checkAccess/checkAccess';
+import { contentAccessChannel } from '../AccessChannels/contentAccessChannel';
 
 import { injectedSignRawChannel } from './injectedSignRawChannel';
 import { SignRawPopupInput, SignRawPopupOutput } from './types';
@@ -11,7 +11,7 @@ export const contentSignRawChannel = new BrowserChannel<
 
 export function initContentSignRawChannel(origin: string): () => void {
   return injectedSignRawChannel.produce(async (input) => {
-    await checkAccess(input.dAppName, origin);
+    await contentAccessChannel.get({ dAppName: input.dAppName, origin });
     return contentSignRawChannel.get({ ...input, origin });
   });
 }
