+++ src/views/AuthorizeDApp/AuthorizeDApp.stories.tsx
@@ -2,7 +2,7 @@ import { Meta } from '@storybook/react';
 
 import { PopupTestProvider } from '../../utilities/popups/PopupTestProvider';
 
-import { AccessInput } from '../../dApps/AccessChannels/browserAccessChannels';
+import { AccessInput } from '../../dApps/AccessChannels/types';
 import { paths } from '../paths';
 
 import { AuthorizeDApp } from './AuthorizeDApp';
@@ -13,7 +13,7 @@ export default {
 } as Meta;
 
 const mockAccessData: AccessInput = {
-  name: 'KILT-Sporran',
+  dAppName: 'KILT-Sporran',
   origin: 'https://polkadot.js.org/apps/',
 };
