+++ src/views/AuthorizeDApp/AuthorizeDApp.test.tsx
@@ -1,13 +1,13 @@
 import { PopupTestProvider } from '../../utilities/popups/PopupTestProvider';
 import { render } from '../../testing/testing';
 
-import { AccessInput } from '../../dApps/AccessChannels/browserAccessChannels';
+import { AccessInput } from '../../dApps/AccessChannels/types';
 import { paths } from '../paths';
 
 import { AuthorizeDApp } from './AuthorizeDApp';
 
 const mockAccessData: AccessInput = {
-  name: 'KILT-Sporran',
+  dAppName: 'KILT-Sporran',
   origin: 'https://polkadot.js.org/apps/',
 };
